Compute normalized power difference
Previously this code was producing an absolute power difference, this
creates an issue where depending on the amount of active validtors and
inflation in the system the change between any two blocks may always be
greater than the 5% threshold for producing a new validator set.

This minor change normalizes the powers, the difference will always
be expressed as a percentage of uint32 max, since all powers in the
Ethereum contract always sum to that value.