Add an event nonce to valset updated events
This patch does somthing that we've tried to avoid for a while. Which is
adding an event nonce to the validator set updated events. So as to make
it possible to relay them into the Cosmos module oracle.

This design change is driven by problems with validator set pruning. In
the worst case validator sets can be produced every single block. But
the worst case on the EVM chain is similarly bad. It could be halted for
several hours or even days.

In order to retain enough validator sets to ensure bridge continuity and
also prune as many validator sets as safely possible we need this
information.