Prune validator sets based on last observed
This code provides the most efficient possible validator set pruning
that we can prove is correct.

By pruning the last observed validator we can keep the number of valsets
in the store to a very low value during normal operation and safely
extend the required storage in the case that the bridge becomes halted
for some reason.